pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gmail-to-notebooklm"
version = "0.1.0"
description = "Convert Gmail emails from a label into Markdown files for Google NotebookLM"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["gmail", "markdown", "notebooklm", "email", "converter"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "google-api-python-client>=2.100.0,<3.0.0",
    "google-auth>=2.25.0,<3.0.0",
    "google-auth-oauthlib>=1.2.0,<2.0.0",
    "google-auth-httplib2>=0.2.0,<1.0.0",
    "beautifulsoup4>=4.12.0,<5.0.0",
    "lxml>=4.9.0,<6.0.0",
    "html2text>=2024.2.0",
    "click>=8.1.0,<9.0.0",
    "python-dateutil>=2.8.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]

[project.scripts]
gmail-to-notebooklm = "gmail_to_notebooklm.main:cli"

[project.urls]
Homepage = "https://github.com/yourusername/gmail-to-notebooklm"
Repository = "https://github.com/yourusername/gmail-to-notebooklm"
Issues = "https://github.com/yourusername/gmail-to-notebooklm/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["gmail_to_notebooklm*"]
exclude = ["tests*", "docs*"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.tox
  | build
  | dist
  | __pycache__
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=gmail_to_notebooklm",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
